refactor(html): create write.py
